LPS-197236 Merge BasePortletRequestWhiteList into InterruptedPortletRequestWhiteListImpl
